+++ lib/src/logic/client/yandex_gpt_api_client.dart
@@ -25,6 +25,11 @@ class YandexGptApiClient implements YandexGptApi {
           catalog: catalog ?? "",
         );
 
+  /// Set new IAM token.
+  void changeToken(String token) {
+    _client.changeToken(token);
+  }
+
   @override
   Future<TextGenerationResponse> generateText(
     TextGenerationRequest request, {